proper: cluster test
The crdt_cluster_test is reimplemented as property test.
Two preconditions is added to the selected command:
 * connect cannot be executed on itself
 * add/remove can be executed on a cluster member
With these conditions random cluster can be build.

The commands had to be wrapped into a sleep, because the async calls
(gen_server:cast) can cause race condition in the evaulation of the
test.